requirements.txt
python-dotenv>=1.0.0
loguru>=0.7.0

# Vector search
chromadb>=0.5.0
lancedb>=0.4.0

# Graph + DB
sqlalchemy>=2.0.0

# Embeddings / Reranking
openai>=1.0.0

# Benchmarking & evaluation
matplotlib>=3.8.0

# Mem0 (mem-zero)
mem0ai>=0.1.0

# LlamaIndex (choose extras you need)
llama-index-core>=0.10.0
llama-index-embeddings-openai>=0.2.0

pytest>=7.4.0
